+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 **Implemented features:**
 
 - `AffineTransform` struct that helps you easily apply transforms like rotation, scaling and translation to a document's coordinate system.
+- Concatenating the transformation matrix of a page + convenience methods for translation, scaling and rotation.
+- Getting the current transformation matrix of a page.
 
 ## [0.2.0](https://github.com/WeirdMath/SwiftyHaru/tree/0.2.0) (2017-06-25)